+++ config.php
@@ -5,6 +5,7 @@ use humhub\modules\space\widgets\HeaderControlsMenu;
 use humhub\modules\space\widgets\Menu;
 use humhub\modules\space\widgets\Sidebar;
 use humhub\modules\user\widgets\ProfileMenu;
+use humhub\modules\user\widgets\ProfileSidebar;
 use humhub\modules\admin\widgets\AdminMenu;
 use humhub\widgets\TopMenu;
 use k7zz\humhub\bbb\Events;
@@ -22,6 +23,7 @@ return [
         ['class' => HeaderControlsMenu::class, 'event' => HeaderControlsMenu::EVENT_INIT, 'callback' => [Events::class, 'onSpaceHeaderControlsMenuInit']],
         ['class' => AdminMenu::class, 'event' => AdminMenu::EVENT_INIT, 'callback' => [Events::class, 'onAdminMenuInit']],
         ['class' => Sidebar::class, 'event' => Sidebar::EVENT_INIT, 'callback' => [Events::class, 'onSpaceSidebarInit']],
+        ['class' => ProfileSidebar::class, 'event' => ProfileSidebar::EVENT_INIT, 'callback' => [Events::class, 'onProfileSidebarInit']],
         ['class' => DashboardSidebar::class, 'event' => DashboardSidebar::EVENT_INIT, 'callback' => [Events::class, 'onDashboardSidebarInit']],
     ],
     'urlManagerRules' => [