+++ models/Session.php
@@ -26,6 +26,7 @@ use k7zz\humhub\bbb\permissions\{
  * @property int    $id
  * @property string $uuid
  * @property string $moderator_pw
+ * @property bool   $notify_on_start
  * ...
  */
 class Session extends ContentActiveRecord
@@ -239,7 +240,7 @@ class Session extends ContentActiveRecord
                 ],
                 'integer'
             ],
-            [['enabled', 'public_join', 'allow_recording', 'show_in_sidebar', 'is_space_default', 'start_chat_minimized', 'start_participants_minimized', 'start_presentation_hidden'], 'boolean'],
+            [['enabled', 'public_join', 'allow_recording', 'show_in_sidebar', 'is_space_default', 'start_chat_minimized', 'start_participants_minimized', 'start_presentation_hidden', 'notify_on_start'], 'boolean'],
             [['public_token'], 'string', 'max' => 64],
             ['layout', 'required'],
             ['layout', 'in', 'range' => Layouts::values()],